+++ Network/Telegram/API/Bot/Object/Callback/Notification.hs
@@ -1,3 +1,0 @@
-module Network.Telegram.API.Bot.Object.Callback.Notification (Notification) where
-
-data Notification