+++ Network/Telegram/API/Bot/Object/Update/Moving.hs
@@ -1,4 +1,4 @@
-module Network.Telegram.API.Bot.Object.Moving (Moving (..)) where
+module Network.Telegram.API.Bot.Object.Update.Moving (Moving (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
 import "aeson" Data.Aeson.Types (Object, Parser)
@@ -7,7 +7,7 @@ import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 
-import Network.Telegram.API.Bot.Object.Message (From)
+import Network.Telegram.API.Bot.Object.Update.Message.From (From)
 
 data Moving = Gone From | Joined [From] deriving Show